+++ src/pages/Authentication.js
@@ -1,7 +1,8 @@
 import { useSearchParams } from 'react-router-dom';
 
-import Signin from '@components/auth/Signin';
-import Signup from '@components/auth/Signup';
+import Signin from '@components/auth/Signin/Signin';
+import Signup from '@components/auth/Signup/Signup';
+import Account from '@components/auth/Account/Account';
 
 const AuthenticationPage = () => {
   const [searchParams] = useSearchParams();
@@ -11,7 +12,7 @@ const AuthenticationPage = () => {
     <>
       {mode === 'signin' && <Signin />}
       {mode === 'signup' && <Signup />}
-      {mode === 'account' && 'ACCOUNT'}
+      {mode === 'account' && <Account />}
     </>
   );
 };
